+++ spotify/__init__.py
@@ -118,6 +118,7 @@ from spotify.eventloop import *  # noqa
 from spotify.image import *  # noqa
 from spotify.inbox import *  # noqa
 from spotify.link import *  # noqa
+from spotify.offline import *  # noqa
 from spotify.playlist import *  # noqa
 from spotify.playlist_container import *  # noqa
 from spotify.playlist_track import *  # noqa